Refine entity relationships with Jackson references

+++ AplicacionFinancieraAPI/src/main/java/com/example/AplicacionFinancieraAPI/Repositories/IUsuarioRepository.java
@@ -4,8 +4,6 @@ import com.example.AplicacionFinancieraAPI.Models.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface IUsuarioRepository extends JpaRepository<Usuario,Integer> {
+public interface IUsuarioRepository extends JpaRepository<Usuario, Long> {
 }
